+++ artifacts/omniproject/src/pages/Settings.tsx
@@ -30,6 +30,7 @@ import { RateGridAdmin } from "../components/settings/RateGridAdmin";
 import { IdentityMapAdmin } from "../components/settings/IdentityMapAdmin";
 import { CostRulesAdmin } from "../components/settings/CostRulesAdmin";
 import { CustomReportsAdmin } from "../components/settings/CustomReportsAdmin";
+import { ContentPagesAdmin } from "../components/settings/ContentPagesAdmin";
 import { GovernanceRulesAdmin } from "../components/settings/GovernanceRulesAdmin";
 import { ScopeUpliftAdmin } from "../components/settings/ScopeUpliftAdmin";
 import { FeatureGovernance } from "../components/settings/FeatureGovernance";
@@ -375,6 +376,10 @@ export function Settings() {
 
       <div className="mt-10">
         <CustomReportsAdmin />
+      </div>
+
+      <div className="mt-10">
+        <ContentPagesAdmin />
       </div>
 
       <div className="mt-10">